excel e filtro

+++ src/pages/listas/experimentos/experimento/index.tsx
@@ -41,7 +41,7 @@ import {
   Input,
 } from '../../../../components';
 import ITabs from '../../../../shared/utils/dropdown';
-import { fetchWrapper } from '../../../../helpers';
+import { tableGlobalFunctions } from '../../../../helpers';
 
 interface IFilter {
   filterFoco: string;
@@ -52,6 +52,7 @@ interface IFilter {
   filterCod: string;
   filterPeriod: string;
   filterDelineamento: string;
+  filterRepetition: string;
   filterRepetitionFrom: string | any;
   filterRepetitionTo: string | any;
   orderBy: object | any;
@@ -60,7 +61,6 @@ interface IFilter {
 
 export interface IExperimento {
   id: number;
-  protocol_name: string;
   experiment_name: string;
   year: number;
   rotulo: string;
@@ -85,6 +85,9 @@ interface IData {
   idSafra: number;
   pageBeforeEdit: string | any;
   filterBeforeEdit: string | any;
+  typeOrderServer :any| string;
+  orderByserver : any |string;
+  cultureId :number | any;
 }
 
 export default function Listagem({
@@ -95,6 +98,9 @@ export default function Listagem({
   idSafra,
   pageBeforeEdit,
   filterBeforeEdit,
+  typeOrderServer,
+  orderByserver,
+  cultureId,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) {
   const { TabsDropDowns } = ITabs;
 
@@ -108,7 +114,7 @@ export default function Listagem({
   const preferences = userLogado.preferences.experimento || {
     id: 0,
     table_preferences:
-      'id,protocolName,foco,type_assay,gli,experimentName,tecnologia,period,delineamento,repetitionsNumber,status,action',
+      'id,foco,type_assay,gli,experimentName,tecnologia,period,delineamento,repetitionsNumber,status,action',
   };
   const [camposGerenciados, setCamposGerenciados] = useState<any>(
     preferences.table_preferences,
@@ -125,7 +131,6 @@ export default function Listagem({
   const [statusAccordion, setStatusAccordion] = useState<boolean>(false);
   const [generatesProps, setGeneratesProps] = useState<IGenerateProps[]>(() => [
     // { name: 'CamposGerenciados[]', title: 'Favorito', value: 'id' },
-    { name: 'CamposGerenciados[]', title: 'Protocolo', value: 'protocolName' },
     { name: 'CamposGerenciados[]', title: 'Foco', value: 'foco' },
     { name: 'CamposGerenciados[]', title: 'Ensaio', value: 'type_assay' },
     { name: 'CamposGerenciados[]', title: 'GLI', value: 'gli' },
@@ -146,7 +151,7 @@ export default function Listagem({
     { name: 'CamposGerenciados[]', title: 'Ações', value: 'action' },
   ]);
 
-  const [orderBy, setOrderBy] = useState<string>('');
+  // const [orderBy, setOrderBy] = useState<string>('');
   const [orderType, setOrderType] = useState<string>('');
   const [colorStar, setColorStar] = useState<string>('');
   const [order, setOrderParams] = useState<string>('');
@@ -154,6 +159,10 @@ export default function Listagem({
   const take: number = itensPerPage;
   const total: number = itemsTotal <= 0 ? 1 : itemsTotal;
   const pages = Math.ceil(total / take);
+  const [orderBy, setOrderBy] = useState<string>(orderByserver);
+  const [typeOrder, setTypeOrder] = useState<string>(typeOrderServer);
+  const pathExtra = `skip=${currentPage * Number(take)}&take=${take}&orderBy=${orderBy == 'tecnologia' ? 'assay_list.tecnologia.cod_tec' : orderBy}&typeOrder=${typeOrder}`;
+
   const [filtersParams, setFiltersParams] = useState<any>(filterBeforeEdit); // Set filter Parameter
 
   const [statusFilter, setStatusFilter] = useState<IGenerateProps[]>(() => [
@@ -187,18 +196,6 @@ export default function Listagem({
       value: 'ETIQ. FINALIZADA',
       defaultChecked: () => camposGerenciados.includes('ETIQ. FINALIZADA'),
     },
-    {
-      name: 'StatusCheckbox',
-      title: 'PARCIALMENTE ALOCADO',
-      value: 'PARCIALMENTE ALOCADO',
-      defaultChecked: () => camposGerenciados.includes('PARCIALMENTE ALOCADO'),
-    },
-    {
-      name: 'StatusCheckbox',
-      title: 'TOTALMENTE ALOCADO',
-      value: 'TOTALMENTE ALOCADO',
-      defaultChecked: () => camposGerenciados.includes('TOTALMENTE ALOCADO'),
-    },
   ]);
 
   const formik = useFormik<IFilter>({
@@ -226,6 +223,7 @@ export default function Listagem({
       filterCod,
       filterPeriod,
       filterDelineamento,
+      filterRepetition,
       filterRepetitionTo,
       filterRepetitionFrom,
     }) => {
@@ -273,18 +271,9 @@ export default function Listagem({
 
       const parametersFilter = `filterRepetitionTo=${filterRepetitionTo}&filterRepetitionFrom=${filterRepetitionFrom}&filterFoco=${filterFoco}&filterTypeAssay=${filterTypeAssay}&filterGli=${filterGli}&filterExperimentName=${filterExperimentName}&filterTecnologia=${filterTecnologia}&filterPeriod=${filterPeriod}&filterRepetition=${filterRepetition}&filterDelineamento=${filterDelineamento}&idSafra=${idSafra}&filterCod=${filterCod}&filterStatus=${filterStatus}&id_culture=${cultureId}`;
 
-      setFiltersParams(parametersFilter);
       setFilter(parametersFilter);
-      setCookies('filterBeforeEdit', filter);
-
-      await experimentService
-        .getAll(`${parametersFilter}&skip=0&take=${itensPerPage}`)
-        .then((response) => {
-          setFilter(parametersFilter);
-          setExperimento(response.response);
-          setTotalItems(response.total);
-          setCurrentPage(0);
-        });
+      setCurrentPage(0);
+      await callingApi(parametersFilter);
     },
   });
 
@@ -314,36 +303,46 @@ export default function Listagem({
     column: string,
     order: string | any,
   ): Promise<void> {
-    // Manage orders of colunms
-    const parametersFilter = await fetchWrapper.handleOrderGlobal(
-      column,
-      order,
-      filter,
-      'experimento',
-    );
+    // // Manage orders of colunms
+    // const parametersFilter = await fetchWrapper.handleOrderGlobal(
+    //   column,
+    //   order,
+    //   filter,
+    //   'experimento',
+    // );
 
-    setOrderParams(parametersFilter);
+    // setOrderParams(parametersFilter);
 
-    const value = await fetchWrapper.skip(currentPage, parametersFilter);
+    // const value = await fetchWrapper.skip(currentPage, parametersFilter);
 
-    await experimentService.getAll(value).then(({ status, response }: any) => {
-      if (status === 200) {
-        setExperimento(response);
-        setFiltersParams(parametersFilter);
-      }
-    });
+    // await experimentService.getAll(value).then(({ status, response }: any) => {
+    //   if (status === 200) {
+    //     setExperimento(response);
+    //     setFiltersParams(parametersFilter);
+    //   }
+    // });
 
-    if (orderList === 2) {
-      setOrder(0);
-      setArrowOrder(<AiOutlineArrowDown />);
-    } else {
-      setOrder(orderList + 1);
-      if (orderList === 1) {
-        setArrowOrder(<AiOutlineArrowUp />);
-      } else {
-        setArrowOrder('');
-      }
-    }
+    // if (orderList === 2) {
+    //   setOrder(0);
+    //   setArrowOrder(<AiOutlineArrowDown />);
+    // } else {
+    //   setOrder(orderList + 1);
+    //   if (orderList === 1) {
+    //     setArrowOrder(<AiOutlineArrowUp />);
+    //   } else {
+    //     setArrowOrder('');
+    //   }
+    // }
+
+    // Gobal manage orders
+    const {
+      typeOrderG, columnG, orderByG, arrowOrder,
+    } = await tableGlobalFunctions.handleOrderG(column, order, orderList);
+
+    setTypeOrder(typeOrderG);
+    setOrderBy(columnG);
+    setOrder(orderByG);
+    setArrowOrder(arrowOrder);
   }
 
   function headerTableFactory(name: any, title: string) {
@@ -407,7 +406,8 @@ export default function Listagem({
   // }
 
   async function deleteItem(id: number) {
-    const { status, message } = await await experimentService.deleted(id);
+    // eslint-disable-next-line max-len
+    const { status, message } = await await experimentService.deleted({ id, userId: userLogado.id });
     if (status === 200) {
       router.reload();
     } else {
@@ -433,9 +433,10 @@ export default function Listagem({
               onClick={() => {
                 setCookies('pageBeforeEdit', currentPage?.toString());
                 setCookies('filterBeforeEdit', filter);
-                localStorage.setItem('filterValueEdit', filtersParams);
-                localStorage.setItem('pageBeforeEdit', currentPage?.toString());
-                // localStorage.setItem('orderSorting', order);
+                setCookies('filterBeforeEditTypeOrder', typeOrder);
+                setCookies('filterBeforeEditOrderBy', orderBy);
+                setCookies('filtersParams', filtersParams);
+                setCookies('lastPage', 'atualizar');
                 router.push(
                   `/listas/experimentos/experimento/atualizar?id=${rowData.id}`,
                 );
@@ -490,11 +491,6 @@ export default function Listagem({
     const tableFields: any = [];
 
     Object.keys(columnCampos).forEach((_, index) => {
-      if (columnCampos[index] === 'protocolName') {
-        tableFields.push(
-          headerTableFactory('Protocolo', 'assay_list.protocol_name'),
-        );
-      }
       if (columnCampos[index] === 'foco') {
         tableFields.push(headerTableFactory('Foco', 'assay_list.foco.name'));
       }
@@ -593,11 +589,10 @@ export default function Listagem({
 
   const downloadExcel = async (): Promise<void> => {
     await experimentService
-      .getAll(filtersParams)
+      .getAll(filter)
       .then(({ status, response, message }: any) => {
         if (status === 200) {
           response.map((item: any) => {
-            console.log(response);
             const newItem = item;
             newItem.SAFRA = item.assay_list?.safra?.safraName;
             newItem.FOCO = item.assay_list?.foco.name;
@@ -618,7 +613,6 @@ export default function Listagem({
             newItem.CLP = item?.clp;
             newItem.EEL = item?.eel;
             newItem.OBSERVAÇÕES = item?.comments;
-            newItem.PROTOCOLO = item.assay_list?.protocol_name;
             newItem.COUNT_NT = newItem.countNT;
             newItem.NPE_QT = newItem.npeQT;
 
@@ -665,36 +659,32 @@ export default function Listagem({
       });
   };
 
-  function handleTotalPages(): void {
+  // manage total pages
+  async function handleTotalPages() {
     if (currentPage < 0) {
       setCurrentPage(0);
     }
   }
 
   async function handlePagination(): Promise<void> {
-    // manage using comman function
-    const { parametersFilter, currentPages } = await fetchWrapper.handlePaginationGlobal(currentPage, take, filtersParams);
+    // // manage using comman function
+    // const { parametersFilter, currentPages } = await fetchWrapper.handlePaginationGlobal(currentPage, take, filtersParams);
 
-    await experimentService.getAll(`${parametersFilter}&idSafra=${idSafra}`).then(({ status, response }: any) => {
-      if (status === 200) {
-        setExperimento(response);
-        // setFiltersParams(parametersFilter);
-        // setTotalItems(response.total); //Set new total records
-        // setCurrentPage(currentPages); //Set new current page
-        setTimeout(removestate, 10000); // Remove State
-      }
-    });
+    // await experimentService.getAll(`${parametersFilter}&idSafra=${idSafra}`).then(({ status, response }: any) => {
+    //   if (status === 200) {
+    //     setExperimento(response);
+    //     // setFiltersParams(parametersFilter);
+    //     // setTotalItems(response.total); //Set new total records
+    //     // setCurrentPage(currentPages); //Set new current page
+    //     setTimeout(removestate, 10000); // Remove State
+    //   }
+    // });
+    await callingApi(filter); // handle pagination globly
   }
 
-  // remove states
-  function removestate() {
-    localStorage.removeItem('filterValueEdit');
-    localStorage.removeItem('pageBeforeEdit');
-  }
-
-  // Checkingdefualt values
+  // Checking defualt values
   function checkValue(value: any) {
-    const parameter = fetchWrapper.getValueParams(value);
+    const parameter = tableGlobalFunctions.getValuesForFilter(value, filtersParams);
     return parameter;
   }
 
@@ -1062,13 +1052,17 @@ export const getServerSideProps: GetServerSideProps = async ({
   )?.response[0]?.itens_per_page) ?? 10;
 
   const { token } = req.cookies;
+  const { cultureId } = req.cookies;
+
   const idSafra = Number(req.cookies.safraId);
   const pageBeforeEdit = req.cookies.pageBeforeEdit
     ? req.cookies.pageBeforeEdit
     : 0;
+
   const filterBeforeEdit = req.cookies.filterBeforeEdit
     ? req.cookies.filterBeforeEdit
-    : '';
+    : `idSafra=${idSafra}&id_culture=${cultureId}`;
+
   const filterApplication = req.cookies.filterBeforeEdit
     ? `${req.cookies.filterBeforeEdit}`
     : `idSafra=${idSafra}&id_culture=${cultureId}`;
@@ -1096,11 +1090,14 @@ export const getServerSideProps: GetServerSideProps = async ({
 
   removeCookies('filterBeforeEdit', { req, res });
   removeCookies('pageBeforeEdit', { req, res });
+  removeCookies('filterBeforeEditTypeOrder', { req, res });
+  removeCookies('filterBeforeEditOrderBy', { req, res });
+  removeCookies('lastPage', { req, res });
 
   const { publicRuntimeConfig } = getConfig();
   const baseUrl = `${publicRuntimeConfig.apiUrl}/experiment`;
 
-  const param = `skip=0&take=${itensPerPage}&idSafra=${idSafra}`;
+  const param = `skip=0&take=${itensPerPage}&idSafra=${idSafra}&id_culture=${cultureId}`;
 
   const urlParameters: any = new URL(baseUrl);
   urlParameters.search = new URLSearchParams(param).toString();
@@ -1123,6 +1120,9 @@ export const getServerSideProps: GetServerSideProps = async ({
       idSafra,
       pageBeforeEdit,
       filterBeforeEdit,
+      orderByserver,
+      typeOrderServer,
+      cultureId,
     },
   };
 };
